// apps/api-functions/src/domain/enums/FormType.ts
/**
 * @fileoverview FormType - Enum for contact manager form types
 * @description Defines the possible types of contact manager forms
 */

/**
 * Enum representing the types of contact manager forms
 */
export enum FormType {
  /** Disconnections form type */
  DISCONNECTIONS = 'Disconnections',
  /** Admissions form type */
  ADMISSIONS = 'Admissions',
  /** Assistance form type */
  ASSISTANCE = 'Assistance'
}